Begun reorganization for utils and gui
