+++ gui/frames/forecast_request_info.py
@@ -5,7 +5,7 @@ from gui.empyrean.labelframe import LabelFrame
 
 from utils.WidgetEnum import WidgetType
 from utils.gridplacement import GridPlacement
-from utils.json.location import Location
+from utils.forecast.location import Location
 
 
 class ForecastRequestInfo_LabelFrame(LabelFrame):
