--- utils/dataclasses/forecast_type.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-class ForecastType(Enum):
-    POINTS      = 'points'
-    HOURLY      = 'hourly'
-    EXTENDED    = 'extended'
-
-    def get_list_of_types(self) -> list:
-        return [ForecastType.POINTS, ForecastType.HOURLY, ForecastType.EXTENDED]
